Finalized R
