+++ riscv_assembler/new.py
@@ -22,9 +22,9 @@ from parse import *
 class AssemblyConverter:
 
 	def __init__(self, output_mode = 'b', nibble_mode = False, hex_mode = False):
-		self.output_mode = __check_output_mode(output_mode)
-		self.nibble_mode = __check_nibble_mode(nibble_mode)
-		self.hex_mode = __check_hex_mode(hex_mode)
+		self.output_mode = self.__check_output_mode(output_mode)
+		self.nibble_mode = self.__check_nibble_mode(nibble_mode)
+		self.hex_mode = self.__check_hex_mode(hex_mode)
 
 	def __str__(self):
 		return "Output: {output_mode}, Nibble: {nibble_mode}, Hex: {hex_mode}".format(
@@ -47,55 +47,61 @@ class AssemblyConverter:
 			hex_mode = self.hex_mode
 		)
 
-	@staticmethod
-	def __check_output_mode(x):
+	def __check_output_mode(self, x):
 		assert x in ['b', 't', 'p', None], "Output Mode needs to be one of b(inary), t(ext), p(rint), or None."
 		return x
 
-	@staticmethod
-	def __check_nibble_mode(x):
+	def __check_nibble_mode(self, x):
 		assert type(x) == bool, "Nibble mode needs to be a boolean."
 		return x
 
-	@staticmethod
-	def __check_hex_mode(x):
+	def __check_hex_mode(self, x):
 		assert type(x) == bool, "Hex mode needs to be a boolean."
 		return x
-
-	'''
-		Property: whether to return as hex or not
-			True = hex
-			False = binary
-	'''
-	@property
-	def hex_mode(self):
-		return self.hex_mode
 
 	'''
 		Property: the way to output machine code
 			Options: 'b', 't', 'p'
 	'''
-	@property
-	def output_mode(self):
-		return self.output_mode
+	'''@property
+				def output_mode(self):
+					return self.output_mode
+			
+				@output_mode.setter
+				def output_mode(self, x):
+					self.output_mode = x'''
 
 	'''
 		Property: whether to print in nibbles (only applicable for text or print)
 			True = nibble
 			False = full number
 	'''
-	@property
-	def nibble_mode(self):
-		return self.nibble_mode
+	'''@property
+				def nibble_mode(self):
+					return self.nibble_mode
+			
+				@nibble_mode.setter
+				def nibble_mode(self, x):
+					self.nibble_mode = x'''
 
-	@staticmethod
-	def parse(file):
-		...
+	'''
+	Property: whether to return as hex or not
+		True = hex
+		False = binary
+	'''
+	'''@property
+				def hex_mode(self):
+					return self.hex_mode
+			
+				@hex_mode.setter
+				def hex_mode(self, x):
+					self.hex_mode = x'''
 
 	'''
 		Put it all together
 	'''
 	def convert(self, file):
-		...
+		result = read_file(file)
+		return result
